pyproject.toml
[build-system]
requires = [
    "setuptools >= 61.0",
    "wheel",
    "setuptools_scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "seqnado"
authors = [
    {name = "Alastair Smith", email = "alastair.smith@ndcls.ox.ac.uk"},
    {name = "Catherine Chahrour", email = "catherine.chahrour@msdtc.ox.ac.uk"}
]
description = "Pipelines for genomics analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
dynamic = ["version"]
dependencies = [
    "click<=8.3.2",
    "cookiecutter<=2.7.1",
    "numpy>=1.24,<=2.2.6",
    "pandas>=2.0,<=2.3.3",
    "pandera<=0.31.1",
    "pexpect<=4.9.0",
    "pip",
    "pulp<=3.3.0",
    "pybtex",
    "pydantic<=2.13.2",
    "pyyaml<=6.0.3",
    "rich<=15.0.0",
    "seaborn<=0.13.2",
    "snakemake-wrapper-utils<=0.8.0",
    "snakemake>=9.12.0,<=9.14.5",
    "tracknado>=0.3.1,<1.0.0",
    "typer<=0.24.1",
    "wget<=3.2",
]

[project.optional-dependencies]
slurm = ["snakemake-executor-plugin-slurm<=2.0.3"]
aws = ["snakemake-executor-plugin-aws-batch"]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings>=1.0.0",
    "mkdocstrings-python>=2.0.1",
    "mkdocs-typer2>=0.1.6",
    "mkdocs-gen-files>=0.6.0",
    "mkdocs-jupyter>=0.25.1",
    "mkdocs-autorefs>=1.4.3",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-section-index>=0.3.10",
    "pygments>=2.19.2",
]

[project.scripts]
seqnado = "seqnado.cli:app"

[project.urls]
Homepage = "https://Milne-Group.github.io/SeqNado"

[tool.setuptools_scm]
write_to = "seqnado/_version.py"
local_scheme = "no-local-version"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests", "tests.*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, multiple components)",
    "edge_case: Edge case tests",
    "slow: Slow tests (may take significant time)",
    "requires_data: Tests that require test data files",
    "parametrize: Parametrized tests",
    "pipeline: End-to-end Snakemake pipeline tests (opt-in)",
    "snakemake: Tests that invoke Snakemake via subprocess",
    "requires_apptainer: Tests that require Apptainer/Singularity",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["seqnado"]
omit = [
    "seqnado/_version.py",
    "seqnado/workflow/scripts/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if typing.TYPE_CHECKING:",
    "@(abc\\.)?abstractmethod",
]
